backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==3.2.16
django-colorfield==0.11.0
django-filter==23.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
environs==11.0.0
gunicorn==20.1.0
marshmallow==3.21.1
mccabe==0.7.0
packaging==24.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.1.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
webcolors==1.11.1
